do not explicitly pin "awscli-cwlogs" version, refs #4
This is to prevent version conflicts in applications that use this
library and might require different versions of transitive
dependencies of "awscli-cwlogs". The previously pinned version of
"awscli-cwlogs" (1.4.0) specifially pinned "awscli" to 1.9.8, which
caused issues. Current version of "awscli-cwlogs" (1.4.4) does not pin
"awscli" anymore.

TODO: not sure how much AWS changes internal library APIs, we
might be able to loosen the upper version restriction